requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
flask_jwt_extended==4.6.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
